Redo the MIG sampling, more robust, interleave data in memory.
The interleaving leaves the samples in DDR2 memory in the
same order as they came in from the ADC which ought to make
code that postprocesses the samples a bit easier to read.

This way the samples come in the correc order in DDR2 memory which ought
to make code that wants to postprocess the data easier to write.